collective indexing into Vertex- and EdgeData

+++ src/NetworkStructures.jl
@@ -167,12 +167,12 @@ struct EdgeData{GDB,elE} <: AbstractVector{elE}
     len::Int
 end
 
-Base.@propagate_inbounds function getindex(e_dat::EdgeData, idx)
-    e_dat.gdb.e_array[idx+e_dat.idx_offset]
+Base.@propagate_inbounds function getindex(e_dat::EdgeData, idx::Int)
+    e_dat.gdb.e_array[idx + e_dat.idx_offset]
 end
 
-Base.@propagate_inbounds function setindex!(e_dat::EdgeData, x, idx)
-    e_dat.gdb.e_array[idx+e_dat.idx_offset] = x
+Base.@propagate_inbounds function setindex!(e_dat::EdgeData, x, idx::Int)
+    e_dat.gdb.e_array[idx + e_dat.idx_offset] = x
     nothing
 end
 
@@ -205,12 +205,12 @@ struct VertexData{GDB,elV} <: AbstractVector{elV}
     len::Int
 end
 
-Base.@propagate_inbounds function getindex(v_dat::VertexData, idx)
-    v_dat.gdb.v_array[idx+v_dat.idx_offset]
+Base.@propagate_inbounds function getindex(v_dat::VertexData, idx::Int)
+    v_dat.gdb.v_array[idx + v_dat.idx_offset]
 end
 
-Base.@propagate_inbounds function setindex!(v_dat::VertexData, x, idx)
-    v_dat.gdb.v_array[idx+v_dat.idx_offset] = x
+Base.@propagate_inbounds function setindex!(v_dat::VertexData, x, idx::Int)
+    v_dat.gdb.v_array[idx + v_dat.idx_offset] = x
     nothing
 end
